+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 [Release Section](https://github.com/hoxfon/react-native-twilio-programmable-voice/releases)
 
+## 0.3.0
+
+- Check if Google Play Services are available before initialising Twilio for receiving calls.
+- Method initWithToken returns a Promise to let the application know if the initialisation did succeed.
+
+
 ## 0.2.2
 
 - fix the instruction to setup the `AndroidManifest.xml` file